+++ published/kinetic_request_ce_bridge_update_v1/handler/init.rb
@@ -24,20 +24,27 @@ class KineticRequestCeBridgeUpdateV1
   end
 
   def execute
+    space_slug = @parameters["space_slug"].empty? ? @info_values["space_slug"] : @parameters["space_slug"]
+    if @info_values['api_server'].include?("${space}")
+      server = @info_values['api_server'].gsub("${space}", space_slug)
+    elsif !space_slug.to_s.empty?
+      server = @info_values['api_server']+"/"+space_slug
+    else
+      server = @info_values['api_server']
+    end
+
     begin
       # Initialize variables
-      space_slug = @parameters["space_slug"].empty? ? @info_values["space_slug"] : @parameters["space_slug"]
       error_handling  = @parameters["error_handling"]
       error_message = nil
 
       api_username    = URI.encode(@info_values["api_username"])
       api_password    = @info_values["api_password"]
-      api_server      = @info_values["api_server"]
       original_name   = URI.escape(@parameters["original_bridge_name"])
 
       raise StandardError, "The original bridge name is required." if original_name.empty?
 
-      api_route = "#{api_server}/#{space_slug}/app/api/v1/bridges/#{original_name}"
+      api_route = "#{server}/app/api/v1/bridges/#{original_name}"
 
       resource = RestClient::Resource.new(api_route, { :user => api_username, :password => api_password })