+++ published/kinetic_request_ce_datastore_submission_create_v1/handler/init.rb
@@ -25,16 +25,23 @@ class KineticRequestCeDatastoreSubmissionCreateV1
 
   def execute
     space_slug = @parameters["space_slug"].empty? ? @info_values["space_slug"] : @parameters["space_slug"]
+    if @info_values['api_server'].include?("${space}")
+      server = @info_values['api_server'].gsub("${space}", space_slug)
+    elsif !space_slug.to_s.empty?
+      server = @info_values['api_server']+"/"+space_slug
+    else
+      server = @info_values['api_server']
+    end
+
     error_handling  = @parameters["error_handling"]
     error_message = nil
 
     api_username    = URI.encode(@info_values["api_username"])
     api_password    = @info_values["api_password"]
-    api_server      = @info_values["api_server"]
     form_slug       = @parameters["form_slug"]
 
     begin
-      api_route = "#{api_server}/#{space_slug}/app/api/v1/datastore/forms/#{form_slug}/submissions"
+      api_route = "#{server}/app/api/v1/datastore/forms/#{form_slug}/submissions"
 
       puts "API ROUTE: #{api_route}" if @enable_debug_logging