+++ published/kinetic_request_ce_datastore_submission_search_v1/handler/init.rb
@@ -28,19 +28,26 @@ class KineticRequestCeDatastoreSubmissionSearchV1
 
   def execute
     space_slug = @parameters["space_slug"].empty? ? @info_values["space_slug"] : @parameters["space_slug"]
+    if @info_values['api_server'].include?("${space}")
+      server = @info_values['api_server'].gsub("${space}", space_slug)
+    elsif !space_slug.to_s.empty?
+      server = @info_values['api_server']+"/"+space_slug
+    else
+      server = @info_values['api_server']
+    end
+
     error_handling  = @parameters["error_handling"]
     error_message = nil
 
     api_username    = URI.encode(@info_values["api_username"])
     api_password    = @info_values["api_password"]
-    api_server      = @info_values["api_server"]
     form_slug      = @parameters["form_slug"]
     query          = @parameters["query"]
 
     count = nil
     formatted_submissions = nil
     begin
-      api_route = "#{api_server}/#{space_slug}/app/api/v1/datastore/forms/#{form_slug}"
+      api_route = "#{server}/app/api/v1/datastore/forms/#{form_slug}"
       api_route += "/submissions?include=details,form,values"
       api_route += "&index=#{URI.encode(@parameters['index'])}" if !@parameters['index'].to_s.empty?
       api_route += "&q=#{CGI.escape(@parameters['query'])}" if !@parameters['query'].to_s.empty?