+++ published/kinetic_request_ce_kapp_update_v1/handler/init.rb
@@ -28,13 +28,20 @@ class KineticRequestCeKappUpdateV1
 
   def execute
     space_slug = @parameters["space_slug"].empty? ? @info_values["space_slug"] : @parameters["space_slug"]
+    if @info_values['api_server'].include?("${space}")
+      server = @info_values['api_server'].gsub("${space}", space_slug)
+    elsif !space_slug.to_s.empty?
+      server = @info_values['api_server']+"/"+space_slug
+    else
+      server = @info_values['api_server']
+    end
+
     error_handling  = @parameters["error_handling"]
     error_message = nil
 
     begin
       # API Route
-      api_route = @info_values["api_server"] +
-                  "/" + space_slug + "/app/api/v1/kapps/" +
+      api_route = server + "/app/api/v1/kapps/" +
                   @parameters["orig_kapp_slug"] + "?include=attributes"
 
       puts "API ROUTE: #{api_route}" if @enable_debug_logging