+++ published/kinetic_request_ce_submission_activity_retrieve_v1/handler/init.rb
@@ -5,7 +5,7 @@ class KineticRequestCeSubmissionActivityRetrieveV1
   def initialize(input)
     # Set the input document attribute
     @input_document = REXML::Document.new(input)
-    
+
     # Retrieve all of the handler info values and store them in a hash variable named @info_values.
     @info_values = {}
     REXML::XPath.each(@input_document, "/handler/infos/info") do |item|
@@ -24,15 +24,22 @@ class KineticRequestCeSubmissionActivityRetrieveV1
   end
 
   def execute
+    space_slug = @parameters["space_slug"].empty? ? @info_values["space_slug"] : @parameters["space_slug"]
+    if @info_values['api_server'].include?("${space}")
+      server = @info_values['api_server'].gsub("${space}", space_slug)
+    elsif !space_slug.to_s.empty?
+      server = @info_values['api_server']+"/"+space_slug
+    else
+      server = @info_values['api_server']
+    end
+
     api_username    = URI.encode(@info_values["api_username"])
     api_password    = @info_values["api_password"]
-    api_server      = @info_values["api_server"]
     submission_id       = @parameters["submission_id"]
     activity_id       = @parameters["activity_id"]
-    space_slug      = @parameters["space_slug"].empty? ? @info_values["space_slug"] : @parameters["space_slug"]
     error_handling  = @parameters["error_handling"]
 
-    api_route = "#{api_server}/#{space_slug}/app/api/v1/submissions/#{submission_id}/activities/#{activity_id}"
+    api_route = "#{server}/app/api/v1/submissions/#{submission_id}/activities/#{activity_id}"
 
     resource = RestClient::Resource.new(api_route, { :user => api_username, :password => api_password })