+++ published/kinetic_request_ce_submission_get_answers_v1/handler/init.rb
@@ -18,7 +18,7 @@ class KineticRequestCeSubmissionGetAnswersV1
   def initialize(input)
     # Set the input document attribute
     @input_document = REXML::Document.new(input)
-    
+
     # Retrieve all of the handler info values and store them in a hash variable named @info_values.
     @info_values = {}
     REXML::XPath.each(@input_document, "/handler/infos/info") do |item|
@@ -37,15 +37,22 @@ class KineticRequestCeSubmissionGetAnswersV1
   # If it returns a result, it will be in a special XML format that the task engine expects. These
   # results will then be available to subsequent tasks in the process.
   def execute
+    space_slug = @parameters["space_slug"].empty? ? @info_values["space_slug"] : @parameters["space_slug"]
+    if @info_values['api_server'].include?("${space}")
+      server = @info_values['api_server'].gsub("${space}", space_slug)
+    elsif !space_slug.to_s.empty?
+      server = @info_values['api_server']+"/"+space_slug
+    else
+      server = @info_values['api_server']
+    end
+
     api_username    = URI.encode(@info_values["api_username"])
     api_password    = @info_values["api_password"]
-    api_server      = @info_values["api_server"]
     kapp_slug       = @parameters["kapp_slug"]
-    space_slug      = @parameters["space_slug"].empty? ? @info_values["space_slug"] : @parameters["space_slug"]
     submission_id   = @parameters["submission_id"]
     error_handling  = @parameters["error_handling"]
 
-    api_route = "#{api_server}/#{space_slug}/app/api/v1/submissions/#{submission_id}?include=values"
+    api_route = "#{server}/app/api/v1/submissions/#{submission_id}?include=values"
 
     puts "API ROUTE: #{api_route}"
 
@@ -55,7 +62,7 @@ class KineticRequestCeSubmissionGetAnswersV1
 
     submission = JSON.parse(response)
 
-    values = submission["submission"]["values"]       
+    values = submission["submission"]["values"]
 
     # Build the results to be returned by this handler
     results = "<results><result name='Handler Error Message'></result>"