+++ published/kinetic_request_ce_submission_update_observers_v1/handler/init.rb
@@ -25,19 +25,26 @@ class KineticRequestCeSubmissionUpdateObserversV1
 
   def execute
     space_slug = @parameters["space_slug"].empty? ? @info_values["space_slug"] : @parameters["space_slug"]
+    if @info_values['api_server'].include?("${space}")
+      server = @info_values['api_server'].gsub("${space}", space_slug)
+    elsif !space_slug.to_s.empty?
+      server = @info_values['api_server']+"/"+space_slug
+    else
+      server = @info_values['api_server']
+    end
+
     error_handling  = @parameters["error_handling"]
     error_message = nil
 
     api_username     = URI.encode(@info_values["api_username"])
     api_password     = @info_values["api_password"]
-    api_server       = @info_values["api_server"]
     submission_id    = @parameters["submission_id"]
     replace_existing = @parameters["replace_existing"].downcase || "false"
     requires_update  = false
     values = {}
 
     begin
-      api_route = "#{api_server}/#{space_slug}/app/api/v1/submissions/#{submission_id}"
+      api_route = "#{server}/app/api/v1/submissions/#{submission_id}"
 
       # Build Resource to get submission values and form fields
       resource = RestClient::Resource.new("#{api_route}?include=values,form.fields", { :user => api_username, :password => api_password })