+++ published/kinetic_request_ce_user_create_v1/handler/init.rb
@@ -25,12 +25,19 @@ class KineticRequestCeUserCreateV1
 
   def execute
     space_slug = @parameters["space_slug"].empty? ? @info_values["space_slug"] : @parameters["space_slug"]
+    if @info_values['api_server'].include?("${space}")
+      server = @info_values['api_server'].gsub("${space}", space_slug)
+    elsif !space_slug.to_s.empty?
+      server = @info_values['api_server']+"/"+space_slug
+    else
+      server = @info_values['api_server']
+    end
+
     error_handling  = @parameters["error_handling"]
     error_message = nil
 
     api_username    = URI.encode(@info_values["api_username"])
     api_password    = @info_values["api_password"]
-    api_server      = @info_values["api_server"]
     new_username    = URI.encode(@parameters["username"])
     new_password    = @parameters["password"]
     email           = @parameters["email"]
@@ -39,7 +46,7 @@ class KineticRequestCeUserCreateV1
 
     results = nil
     begin
-      api_route = "#{api_server}/#{space_slug}/app/api/v1/users"
+      api_route = "#{server}/app/api/v1/users"
 
       resource = RestClient::Resource.new(api_route, { :user => api_username, :password => api_password })