+++ unpublished/kinops_legacy/kinetic_request_ce_kapp_submission_search_v1/test/simple_input.rb
@@ -1,0 +1,15 @@
+{
+  'info' => {
+    'api_server' => 'http://localhost:8080/kinetic',
+    'api_username' => 'user@kineticdata.com',
+    'api_password' => 'password',
+    'space_slug' => 'acme'
+  },
+  'parameters' => {
+    'error_handling' => 'Error Message',
+    'space_slug' => '',
+    'kapp_slug' => '',
+    'query' => '',
+    'return_type' => ''
+  }
+}