+++ unpublished/kinops_legacy/kinetic_request_ce_user_retrieve_from_list_v1/handler/init.rb
@@ -1,0 +1,99 @@
+# Require the dependencies file to load the vendor libraries
+require File.expand_path(File.join(File.dirname(__FILE__), "dependencies"))
+
+class KineticRequestCeUserRetrieveFromListV1
+  # Prepare for execution by building Hash objects for necessary values and
+  # validating the present state.  This method sets the following instance
+  # variables:
+  # * @input_document - A REXML::Document object that represents the input Xml.
+  # * @info_values - A Hash of info names to info values.
+  # * @parameters - A Hash of parameter names to parameter values.
+  #
+  # This is a required method that is automatically called by the Kinetic Task
+  # Engine.
+  #
+  # ==== Parameters
+  # * +input+ - The String of Xml that was built by evaluating the node.xml
+  #   handler template.
+  def initialize(input)
+    # Set the input document attribute
+    @input_document = REXML::Document.new(input)
+
+    # Retrieve all of the handler info values and store them in a hash variable named @info_values.
+    @info_values = {}
+    REXML::XPath.each(@input_document, "/handler/infos/info") do |item|
+      @info_values[item.attributes["name"]] = item.text.to_s.strip
+    end
+
+    # Retrieve all of the handler parameters and store them in a hash variable named @parameters.
+    @parameters = {}
+    REXML::XPath.each(@input_document, "/handler/parameters/parameter") do |item|
+      @parameters[item.attributes["name"]] = item.text.to_s.strip
+    end
+  end
+
+  # The execute method gets called by the task engine when the handler's node is processed. It is
+  # responsible for performing whatever action the name indicates.
+  # If it returns a result, it will be in a special XML format that the task engine expects. These
+  # results will then be available to subsequent tasks in the process.
+  def execute
+    space_slug = @parameters["space_slug"].empty? ? @info_values["space_slug"] : @parameters["space_slug"]
+    if @info_values['api_server'].include?("${space}")
+      server = @info_values['api_server'].gsub("${space}", space_slug)
+    elsif !space_slug.to_s.empty?
+      server = @info_values['api_server']+"/"+space_slug
+    else
+      server = @info_values['api_server']
+    end
+
+    api_username    = URI.encode(@info_values["api_username"])
+    api_password    = @info_values["api_password"]
+    usernames       = @parameters["usernames"]
+    user_details    = []
+    users_not_found = []
+
+
+    usernames.split(",").each{|username|
+      username = URI.encode(username)
+      api_route = "#{server}/app/api/v1/users/#{username}?include=details,attributes,profileAttributes"
+
+      begin
+        resource = RestClient::Resource.new(api_route, { :user => api_username, :password => api_password })
+        response = resource.get
+        user_details.push(JSON.parse(response)["user"])
+      rescue RestClient::Exception => error
+        if error.http_code == 404
+          # Do nothing with the error.  Record doesn't exist which is valid in this context.
+          users_not_found.push(username)
+        else
+          raise StandardError, error
+        end
+      end
+    }
+
+    <<-RESULTS
+    <results>
+     <result name="User Details JSON">#{escape(user_details.to_json)}</result>
+     <result name="Users Not Found JSON">#{escape(users_not_found.to_json)}</result>
+    </results>
+    RESULTS
+
+  end
+
+  ##############################################################################
+  # General handler utility functions
+  ##############################################################################
+
+  # This is a template method that is used to escape results values (returned in
+  # execute) that would cause the XML to be invalid.  This method is not
+  # necessary if values do not contain character that have special meaning in
+  # XML (&, ", <, and >), however it is a good practice to use it for all return
+  # variable results in case the value could include one of those characters in
+  # the future.  This method can be copied and reused between handlers.
+  def escape(string)
+    # Globally replace characters based on the ESCAPE_CHARACTERS constant
+    string.to_s.gsub(/[&"><]/) { |special| ESCAPE_CHARACTERS[special] } if string
+  end
+  # This is a ruby constant that is used by the escape method
+  ESCAPE_CHARACTERS = {'&'=>'&amp;', '>'=>'&gt;', '<'=>'&lt;', '"' => '&quot;'}
+end