+++ published/kinetic_request_ce_attachment_temp_link_retrieve_v1/CHANGELOG.md
@@ -1,3 +1,7 @@
+Kinetic Request CE Attachment Temp Link Retrieve V1.2 (2018-05-25)
+* API Server Info Value changed to allow ${space} in the url for subdomain support
+(ie. https://${space}.localhost:8080/kinetic)
+
 Kinetic Request CE Attachment Temp Link Retrieve V1.1 (2017-11-21)
 * Changed name from Attachment Info to more accurately refer to what the handler is used for